requirements.txt
requests==2.25.1
tqdm==4.64.1
matplotlib==3.2.2
numpy==1.21.6
scikit-learn==1.0.2
pandas==1.3.5
--find-links https://download.pytorch.org/whl/cu116/torch-1.13.1%2Bcu116-cp38-cp38-linux_x86_64.whl
torch==1.13.1+cu116
--find-links https://download.pytorch.org/whl/cu116/torchvision-0.14.1%2Bcu116-cp38-cp38-linux_x86_64.whl
torchvision==0.14.1+cu116
flask==2.1.3
opencv-python==4.6.0.66
Pillow==7.1.2